+++ src/assets/project_screenshots/gms/index.ts
@@ -1,0 +1,15 @@
+import customerView from "./customer_view.png";
+import posView from "./pos_view.png";
+import menuEditor from "./menu_editor.png";
+import promoEditor from "./promo_editor.png";
+import analytics from "./analytics.png";
+import notificationsEditor from "./notifications_editor.png";
+
+export const gms = [
+  customerView,
+  posView,
+  menuEditor,
+  promoEditor,
+  analytics,
+  notificationsEditor,
+];